% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dar16C_hiP}
\alias{Dar16C_hiP}
\title{Example data.}
\format{
A data frame with 792 rows and 18 columns:
}
\source{
\url{CIAT: Bean Breeding Program}
}
\usage{
Dar16C_hiP
}
\description{
A dataset containing an alpha lattice design.
}
\keyword{datasets}
