% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desireGain.R
\name{desireFormat}
\alias{desireFormat}
\title{Desire Gain Format}
\usage{
desireFormat(
  n_traits = 1,
  traits = paste0("trait_", 1:n_traits),
  wt = rep(1, length(traits)),
  heritability = rep(0.5, length(traits)),
  sd = rep(1, length(traits)),
  corPhen = diag(1, length(traits)),
  corGen = diag(1, length(traits))
)
}
\arguments{
\item{n_traits}{number of traits}

\item{traits}{name of the traits}

\item{wt}{weights}

\item{heritability}{vector with heritabilities for each trait}

\item{sd}{vector with genotypic standard deviations}

\item{corPhen}{matrix with phenotypic correlation}

\item{corGen}{matrix with genotypic correlation}
}
\value{
data.frame with desire format
}
\description{
Desire Gain Format
}
\examples{
# desireFormat(n_traits = 4)
}
