% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBLUP.R
\name{GBLUPs}
\alias{GBLUPs}
\title{GBLUP and rrBLUP estimation}
\usage{
GBLUPs(
  pheno_data = NULL,
  geno_matrix = NULL,
  genotype = NULL,
  traits = NULL,
  method = c("GBLUP", "rrBLUP", "mix")
)
}
\arguments{
\item{pheno_data}{data.frame with phenotypic data}

\item{geno_matrix}{data.frame with first column identifying the genotype names
and the rest containing marker information in format (-1, 0 , 1).}

\item{traits}{vector with the traits to be analyzed}

\item{method}{"GBLUP", "rrBLUP" or "mix" ("mix" in progress)}

\item{genoype}{string with the genotype name}
}
\value{
List with GBLUPs, variance components and marker effects.
 list(results, var_comp, markers)
}
\description{
GBLUP and rrBLUP estimation
}
\examples{
# GBLUPs(
# pheno_data = NULL,
# geno_matrix = NULL,
# genoype = NULL,
# traits = NULL,
# method = c("GBLUP", "rrBLUP")
# )
}
