% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpATSplus.R
\name{LSD}
\alias{LSD}
\title{LSD SpATS}
\usage{
LSD(model, alpha = 0.05, data.frame = FALSE)
}
\arguments{
\item{model}{an object of class SpATS as produced by SpATS()}

\item{alpha}{Level of risk for the test (0.05 by default)}

\item{data.frame}{whether return only the LSD value or a complete summary
FALSE by default}
}
\value{
data.frame or a lsd value depending on the data.frame argument
}
\description{
LSD SpATS
}
\examples{
# library(SpATS)
# data(wheatdata)
# wheatdata$R <- as.factor(wheatdata$row)
# wheatdata$C <- as.factor(wheatdata$col)
#
# m1 <- SpATS(
#  response = "yield",
#  spatial = ~ PSANOVA(col, row, nseg = c(10, 20), nest.div = 2),
#  genotype = "geno",
#  genotype.as.random = FALSE,
#  fixed = ~ colcode + rowcode,
#  random = ~ R + C,
#  data = wheatdata,
#  control = list(tolerance = 1e-03)
# )
# LSD(m1, data.frame = TRUE)
}
