% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBLUP.R
\name{marker_plot}
\alias{marker_plot}
\title{Marker Plot}
\usage{
marker_plot(
  marker = NULL,
  map = NULL,
  trait_selected = "",
  type = "point",
  legend_size = 15,
  alpha = 1
)
}
\arguments{
\item{marker}{data.frame with marker effects (marker, trait_1, trait_2, ...).}

\item{map}{data.frame with the genetic map (marker, position, chr).}

\item{trait_selected}{vector with traits to plot}

\item{type}{string selecting whether to use points or lines in the geom.
(point by default).}

\item{legend_size}{numeric value to define the size of the theme 
(15 by default).}

\item{alpha}{numeric value between (0, 1) to define the alpha in the geom.}
}
\value{
ggplot object
}
\description{
Marker Plot
}
\examples{
# marker_plot(
# marker = marker_info,
# map = map,
# trait_selected = traits,
# type = "line",
# alpha = 1
# )
}
