% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mod.list}
\alias{mod.list}
\title{A list of datasets}
\format{
a list
}
\source{
\url{CIAT: Bean Breeding Program}
}
\usage{
mod.list
}
\description{
mod.list is an object that pass a dataset to the environment variables
}
\keyword{datasets}
