% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASReml_MET.R
\name{stageMET}
\alias{stageMET}
\title{Routine to perform final step of a two-stage MET analyses}
\usage{
stageMET(
  data = NULL,
  gen = NULL,
  trial = NULL,
  resp = NULL,
  weight = NULL,
  type.gen = "random",
  type.trial = "fixed",
  vc.model = "corh"
)
}
\arguments{
\item{data}{dataframe with all relevant columns for MET analyses.}

\item{gen}{factor name for genotypes (or treatments)}

\item{trial}{factor name for trial (or environment)}

\item{resp}{column name for the response variable to evaluate}

\item{weight}{column name for the weight of response (default = 1)}

\item{type.gen}{model assumption for genotype effects: 'random' or 'fixed' (default = 'random')}

\item{type.trial}{model assumption for trial effects: 'random' or 'fixed' (default = 'fixed')}

\item{vc.model}{variance-covariance model to fit: 'diag', 'corv', 'corh', 'fa1', 'fa2', 
'fa3', 'fa4', 'corgh' (default = 'corh') (only for type.gen =' random')}
}
\value{
Several objects with reports of the MET analysis. 
call:        String with the ASReml-R call used to fit the requested model
mod:         ASReml-R object with all information from the fitted model
predictions: Predictions for all genotypes across all sites, with thier standard
              error and reliability.
}
\description{
\code{stageMET} Performs the genetic analysis of an MET dataset corresponding
to the final step of a two-stage analysis, where input corresponds to meand (or 
adjusted values) of Evaluates and verifies the data originating from several trials with the 
aim of determining levels of: connectivity, variability, etc., which are reported as 
statistics. Only non-NA observations are considered in the reports. Note that trial is 
always considered a fixed effect.
}
\examples{
# Example 1: 

}
\author{
Salvador A. Gezan. VSN International
}
