% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asreml_funct.R
\name{stats.spatial}
\alias{stats.spatial}
\title{Routine to obtain several goodness-of-fit statistics for spatial and non-spatial models}
\usage{
stats.spatial(
  object = NULL,
  checks = FALSE,
  solution = NULL,
  preds = NULL,
  k = 1,
  ctable = NULL
)
}
\arguments{
\item{object}{model fit output for object class asreml. Genotype factor is called 'gen'}

\item{checks}{logical to indicate if checks were included in the model (default = FALSE)}

\item{solution}{data frame with the random solutions from asreml (BLUPs)}

\item{preds}{data frame with the predictions for genotypes (required to calculate Aopt, Dopt)}

\item{k}{value 1 or 2 to be used for the denominator of heritabilty calculation (herit.PEV) (default = 2)}

\item{ctable}{table with genotypes and checks (optional)}
}
\value{
A table with goodness-of-fit statistics for models evaluated. This includes columns: 
number of variance components in the model (n.VC), log-likelihood (logL), Akaike information 
criteria (AIC), Bayesian information criteria (BIC), heritability based on variance components,
heritability based on predictor error variance (heritPEV, only for genotypes 'random'), A-optimality value, and logarithm of 
D-optimality value.
}
\description{
\code{stats.spatial} Internal routine that calculates goodness-of-fit statisitcs for a
specific fitted model provided as input. Statistics obtained are log-likelihood, AIC, 
BIC, heritability based on predictor error variance (heritPEV, only for genotypes 'random'),
A-optimality value, and logarithm of D-optimality value,
}
\examples{
# Example: Pending

}
\author{
Salvador A. Gezan. VSN International
}
