% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dar16C_hiP}
\alias{Dar16C_hiP}
\title{A dataset containing an alpha lattice design.}
\format{
data frame
\describe{
  \item{dataset}{description here }
  \item{unique_identifier}{description here}
  \item{plot}{description here}
  \item{line}{description here}
  \item{rep}{description here}
  \item{treatment}{description here}
  \item{block}{description here}
  \item{row}{description here}
  \item{col}{description here}
  \item{DF}{description here}
  \item{DPM}{description here}
  \item{PLHA}{description here}
  \item{POM_c2}{description here}
  \item{SW100}{description here}
  \item{TSW}{description here}
  \item{YDHA}{description here}
  \item{YDHPL}{description here}
}
}
\source{
CIAT: Bean Breeding Program
}
\usage{
Dar16C_hiP
}
\description{
A dataset containing an alpha lattice design.
}
\keyword{datasets}
