% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBLUP.R
\name{GBLUPs}
\alias{GBLUPs}
\title{GBLUP and rrBLUP estimation}
\usage{
GBLUPs(
  pheno_data = NULL,
  geno_matrix = NULL,
  genotype = NULL,
  traits = NULL,
  method = c("GBLUP", "rrBLUP", "mix")
)
}
\arguments{
\item{pheno_data}{data.frame with phenotypic data}

\item{geno_matrix}{matrix n_gen by n_marker dimension in format (-1, 0 , 1)}

\item{traits}{vector with the traits to be analyzed}

\item{method}{"GBLUP", "rrBLUP" or "mix"}

\item{genoype}{string with the genotype name}
}
\value{
list with results, information shared (pheno - geno),
 variance components (if GBLUP or mix was fitted) and marker effects.
}
\description{
GBLUP and rrBLUP estimation
}
\examples{
# GBLUPs(pheno_data = NULL,
# geno_matrix = NULL,
# genoype = NULL,
# traits = NULL,
# method = c("GBLUP", "rrBLUP", "mix)
# )
}
